add crate to keep gitoxide conversions to prevent duplication
