+++ Cargo.toml
@@ -31,6 +31,7 @@ members = [
     "crates/gitbutler-operating-modes",
     "crates/gitbutler-edit-mode",
     "crates/gitbutler-cherry-pick",
+    "crates/gitbutler-oxidize",
 ]
 resolver = "2"
 
@@ -83,6 +84,7 @@ gitbutler-diff = { path = "crates/gitbutler-diff" }
 gitbutler-operating-modes = { path = "crates/gitbutler-operating-modes" }
 gitbutler-edit-mode = { path = "crates/gitbutler-edit-mode" }
 gitbutler-cherry-pick = { path = "crates/gitbutler-cherry-pick" }
+gitbutler-oxidize = { path = "crates/gitbutler-oxidize" }
 
 [profile.release]
 codegen-units = 1 # Compile crates one after another so the compiler can optimize better
