+++ crates/gitbutler-branch/Cargo.toml
@@ -15,6 +15,7 @@ gitbutler-id.workspace = true
 gitbutler-error.workspace = true
 gitbutler-fs.workspace = true
 gitbutler-diff.workspace = true
+gitbutler-oxidize.workspace = true
 itertools = "0.13"
 toml.workspace = true
 serde = { workspace = true, features = ["std"] }
