+++ crates/gitbutler-repo/Cargo.toml
@@ -35,6 +35,7 @@ gitbutler-time.workspace = true
 gitbutler-commit.workspace = true
 gitbutler-url.workspace = true
 gitbutler-cherry-pick.workspace = true
+gitbutler-oxidize.workspace = true
 uuid.workspace = true
 itertools = "0.13"
